JBIDE-21841 Duplicated code is moved to superclass to be reused.
