+++ plugins/org.jboss.tools.openshift.ui/src/org/jboss/tools/openshift/internal/ui/server/BuildConfigDetailViews.java
@@ -12,16 +12,11 @@ package org.jboss.tools.openshift.internal.ui.server;
 
 import org.eclipse.core.databinding.DataBindingContext;
 import org.eclipse.core.databinding.observable.value.IObservableValue;
-import org.eclipse.jface.layout.GridDataFactory;
 import org.eclipse.jface.layout.GridLayoutFactory;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.custom.StyledText;
-import org.eclipse.swt.events.FocusEvent;
-import org.eclipse.swt.events.FocusListener;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Label;
 import org.jboss.tools.openshift.internal.common.ui.detailviews.AbstractStackedDetailViews;
-import org.jboss.tools.openshift.internal.common.ui.utils.StyledTextUtils;
 
 import com.openshift.restclient.model.IBuildConfig;
 
@@ -69,35 +64,6 @@ public class BuildConfigDetailViews extends AbstractStackedDetailViews {
 			return container;
 		}
 
-		FocusListener fl = new FocusListener() {
-
-			@Override
-			public void focusGained(FocusEvent e) {
-			}
-
-			@Override
-			public void focusLost(FocusEvent e) {
-				if(e.getSource() instanceof StyledText) {
-					((StyledText)e.getSource()).setSelection(0, 0);
-				}
-			}
-
-		};
-
-		private StyledText createLabeledValue(String labelText, Composite container) {
-			Label label = new Label(container, SWT.NONE);
-			label.setText(labelText);
-			GridDataFactory.fillDefaults()
-				.align(SWT.LEFT, SWT.CENTER).applyTo(label);
-			StyledText styledText = new StyledText(container, SWT.READ_ONLY);
-			styledText.setAlwaysShowScrollBars(false);
-			StyledTextUtils.setTransparent(styledText);
-			GridDataFactory.fillDefaults()
-					.align(SWT.LEFT, SWT.CENTER).grab(true, false).applyTo(styledText);
-			styledText.addFocusListener(fl);
-			return styledText;
-		}
-
 		@Override
 		public void onVisible(IObservableValue serviceObservable, DataBindingContext dbc) {
 			Object value = serviceObservable.getValue();
@@ -113,10 +79,6 @@ public class BuildConfigDetailViews extends AbstractStackedDetailViews {
 			labelsText.setText(labels);
 			sourceURIText.setText(buildConfig.getSourceURI());
 			outputRepositoryNameText.setText(buildConfig.getOutputRepositoryName());
-		}
-
-		@Override
-		public void onInVisible(IObservableValue templateObservable, DataBindingContext dbc) {
 		}
 
 		@Override
